+++ ccf-system/src/main/java/com/xyz/modules/biz/rest/ReleasedPersonController.java
@@ -1,21 +1,32 @@
 package com.xyz.modules.biz.rest;
 
 import com.xyz.aop.log.Log;
+import com.xyz.exception.BadRequestException;
 import com.xyz.modules.biz.domain.ReleasedPerson;
 import com.xyz.modules.biz.service.ReleasedPersonService;
 import com.xyz.modules.biz.service.dto.ReleasedPersonQueryCriteria;
+import com.xyz.modules.security.security.JwtUser;
+import com.xyz.modules.system.service.DeptService;
+import com.xyz.modules.system.service.DictService;
+import com.xyz.utils.SecurityUtils;
+import com.xyz.utils.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import io.swagger.annotations.*;
 
+import java.util.List;
+
 /**
  * @author 刘鑫
  * @date 2020-04-10
+ * 特殊人群刑满释放人员 CRUD
  */
 @Api(tags = "ReleasedPerson管理")
 @RestController
@@ -25,12 +36,37 @@ public class ReleasedPersonController {
     @Autowired
     private ReleasedPersonService ReleasedPersonService;
 
+    @Autowired
+    private DictService dictService;
+    @Autowired
+    @Qualifier("jwtUserDetailsService")
+    private UserDetailsService userDetailsService;
+
+    @Autowired
+    private DeptService deptService;
+
     @Log("查询ReleasedPerson")
     @ApiOperation(value = "查询ReleasedPerson")
     @GetMapping(value = "/ReleasedPerson")
     @PreAuthorize("hasAnyRole('ADMIN','RELEASEDPERSON_ALL','RELEASEDPERSON_SELECT')")
     public ResponseEntity getReleasedPersons(ReleasedPersonQueryCriteria criteria, Pageable pageable){
+        JwtUser u = (JwtUser) userDetailsService.loadUserByUsername(SecurityUtils.getUsername());
+        String deptId = u.getDeptDto().getId();
+        List<String> deptCodes = deptService.getDownGradeDeptCodes(deptId);
+        criteria.setCreator(u.getId());
+        criteria.setUnitCode(deptCodes);
         return new ResponseEntity(ReleasedPersonService.queryAll(criteria,pageable),HttpStatus.OK);
+    }
+
+    @Log("查询详情ReleasedPerson")
+    @ApiOperation(value = "查询详情ReleasedPerson")
+    @GetMapping(value = "/ReleasedPerson/{id}")
+    @PreAuthorize("hasAnyRole('ADMIN','RELEASEDPERSON_ALL','RELEASEDPERSON_SELECT')")
+    public ResponseEntity getById(@PathVariable String id){
+        if (StringUtils.isBlank(id)){
+            throw new BadRequestException("主键ID不能为空");
+        }
+        return new ResponseEntity(ReleasedPersonService.findById(id),HttpStatus.OK);
     }
 
     @Log("新增ReleasedPerson")
@@ -46,6 +82,9 @@ public class ReleasedPersonController {
     @PutMapping(value = "/ReleasedPerson")
     @PreAuthorize("hasAnyRole('ADMIN','RELEASEDPERSON_ALL','RELEASEDPERSON_EDIT')")
     public ResponseEntity update(@Validated @RequestBody ReleasedPerson resources){
+        if (StringUtils.isBlank(resources.getReleasedId())){
+            throw new BadRequestException("主键ID不能为空");
+        }
         ReleasedPersonService.update(resources);
         return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
@@ -55,6 +94,9 @@ public class ReleasedPersonController {
     @DeleteMapping(value = "/ReleasedPerson/{releasedId}")
     @PreAuthorize("hasAnyRole('ADMIN','RELEASEDPERSON_ALL','RELEASEDPERSON_DELETE')")
     public ResponseEntity delete(@PathVariable String releasedId){
+        if (StringUtils.isBlank(releasedId)){
+            throw new BadRequestException("主键ID不能为空");
+        }
         ReleasedPersonService.delete(releasedId);
         return new ResponseEntity(HttpStatus.OK);
     }
